feat: TTL bump on active credit lines
