+++ contracts/credit/src/lifecycle.rs
@@ -1,8 +1,45 @@
+// SPDX-License-Identifier: MIT
+
+//! Credit-line lifecycle operations: open, suspend, close, default, reinstate.
+//!
+//! # TTL bumping
+//! Every function that writes a credit-line entry to persistent storage also
+//! calls [`extend_ttl`](soroban_sdk::storage::Persistent::extend_ttl) to keep
+//! the entry live for [`CREDIT_LINE_TTL_EXTEND_TO`] ledgers.  The threshold
+//! [`CREDIT_LINE_TTL_THRESHOLD`] ensures we only pay for the extension when
+//! there is a real risk of expiry.
+
 use crate::auth::{require_admin, require_admin_auth};
 use crate::events::{publish_credit_line_event, CreditLineEvent};
+use crate::storage::{CREDIT_LINE_TTL_EXTEND_TO, CREDIT_LINE_TTL_THRESHOLD};
 use crate::types::{CreditLineData, CreditStatus};
 use soroban_sdk::{symbol_short, Address, Env};
 
+/// Helper: bump the TTL of a borrower's persistent credit-line entry.
+///
+/// Should be called after every read **or** write that constitutes an
+/// "interaction" with the credit line so the ledger entry never silently
+/// expires while the line is still in use.
+fn bump_credit_line_ttl(env: &Env, borrower: &Address) {
+    env.storage()
+        .persistent()
+        .extend_ttl(borrower, CREDIT_LINE_TTL_THRESHOLD, CREDIT_LINE_TTL_EXTEND_TO);
+}
+
+/// Open a new credit line for a borrower (admin only).
+///
+/// # Parameters
+/// - `borrower`: Address of the borrower.
+/// - `credit_limit`: Maximum drawable amount (must be > 0).
+/// - `interest_rate_bps`: Annual interest rate in basis points (0–10 000).
+/// - `risk_score`: Borrower risk score (0–100).
+///
+/// # Panics
+/// - If `credit_limit` ≤ 0, `interest_rate_bps` > 10 000, or `risk_score` > 100.
+/// - If the borrower already has an Active credit line.
+///
+/// # Events
+/// Emits a `("credit", "opened")` [`CreditLineEvent`].
 pub fn open_credit_line(
     env: Env,
     borrower: Address,
@@ -17,7 +54,7 @@ pub fn open_credit_line(
     );
     assert!(risk_score <= 100, "risk_score must be between 0 and 100");
 
-    // Prevent overwriting an existing Active credit line
+    // Prevent overwriting an existing Active credit line.
     if let Some(existing) = env
         .storage()
         .persistent()
@@ -28,6 +65,7 @@ pub fn open_credit_line(
             "borrower already has an active credit line"
         );
     }
+
     let credit_line = CreditLineData {
         borrower: borrower.clone(),
         credit_limit,
@@ -41,6 +79,8 @@ pub fn open_credit_line(
     };
 
     env.storage().persistent().set(&borrower, &credit_line);
+    // Bump TTL: newly opened lines start with a full TTL window.
+    bump_credit_line_ttl(&env, &borrower);
 
     publish_credit_line_event(
         &env,
@@ -56,16 +96,16 @@ pub fn open_credit_line(
     );
 }
 
-/// Suspend a credit line temporarily.
+/// Suspend a credit line temporarily (admin only).
 ///
-/// Called by admin to freeze a borrower's credit line without closing it.
-/// The credit line can be reactivated or closed after suspension.
+/// Only lines in [`CreditStatus::Active`] status may be suspended.
 ///
 /// # Parameters
 /// - `borrower`: The borrower's address.
 ///
 /// # Panics
 /// - If no credit line exists for the given borrower.
+/// - If the credit line is not currently Active.
 ///
 /// # Events
 /// Emits a `("credit", "suspend")` [`CreditLineEvent`].
@@ -83,6 +123,8 @@ pub fn suspend_credit_line(env: Env, borrower: Address) {
 
     credit_line.status = CreditStatus::Suspended;
     env.storage().persistent().set(&borrower, &credit_line);
+    // Bump TTL: interacting with a suspended line keeps it live.
+    bump_credit_line_ttl(&env, &borrower);
 
     publish_credit_line_event(
         &env,
@@ -98,19 +140,20 @@ pub fn suspend_credit_line(env: Env, borrower: Address) {
     );
 }
 
-/// Close a credit line. Callable by admin (force-close) or by borrower when utilization is zero.
-/// Allowed from Active, Suspended, or Defaulted. Idempotent if already Closed.
+/// Close a credit line. Callable by admin (force-close) or by borrower
+/// when utilization is zero. Idempotent if already Closed.
 ///
 /// # Arguments
-/// * `closer` - Address that must have authorized this call. Must be either the contract admin
-///   (can close regardless of utilization) or the borrower (can close only when
-///   `utilized_amount` is zero).
+/// * `closer` - Must be either the contract admin (can close regardless of
+///   utilization) or the borrower (can close only when `utilized_amount` is zero).
 ///
-/// # Errors
-/// * Panics if credit line does not exist, or if `closer` is not admin/borrower, or if
-///   borrower closes while `utilized_amount != 0`.
+/// # Panics
+/// - If the credit line does not exist.
+/// - If `closer` is not the admin or borrower.
+/// - If the borrower attempts to close while `utilized_amount != 0`.
 ///
-/// Emits a CreditLineClosed event.
+/// # Events
+/// Emits a `("credit", "closed")` [`CreditLineEvent`].
 pub fn close_credit_line(env: Env, borrower: Address, closer: Address) {
     closer.require_auth();
 
@@ -137,6 +180,8 @@ pub fn close_credit_line(env: Env, borrower: Address, closer: Address) {
 
     credit_line.status = CreditStatus::Closed;
     env.storage().persistent().set(&borrower, &credit_line);
+    // Bump TTL: keep the closed record live so history is queryable.
+    bump_credit_line_ttl(&env, &borrower);
 
     publish_credit_line_event(
         &env,
@@ -154,10 +199,14 @@ pub fn close_credit_line(env: Env, borrower: Address, closer: Address) {
 
 /// Mark a credit line as defaulted (admin only).
 ///
-/// Call when the line is past due or when an oracle/off-chain signal indicates default.
 /// Transition: Active or Suspended → Defaulted.
-/// After this, draw_credit is disabled and repay_credit remains allowed.
-/// Emits a CreditLineDefaulted event.
+/// After this, `draw_credit` is disabled and `repay_credit` remains allowed.
+///
+/// # Panics
+/// - If no credit line exists for the given borrower.
+///
+/// # Events
+/// Emits a `("credit", "default")` [`CreditLineEvent`].
 pub fn default_credit_line(env: Env, borrower: Address) {
     require_admin_auth(&env);
     let mut credit_line: CreditLineData = env
@@ -168,6 +217,8 @@ pub fn default_credit_line(env: Env, borrower: Address) {
 
     credit_line.status = CreditStatus::Defaulted;
     env.storage().persistent().set(&borrower, &credit_line);
+    // Bump TTL: defaulted lines must remain queryable during workout period.
+    bump_credit_line_ttl(&env, &borrower);
 
     publish_credit_line_event(
         &env,
@@ -186,6 +237,13 @@ pub fn default_credit_line(env: Env, borrower: Address) {
 /// Reinstate a defaulted credit line to Active (admin only).
 ///
 /// Allowed only when status is Defaulted. Transition: Defaulted → Active.
+///
+/// # Panics
+/// - If no credit line exists for the given borrower.
+/// - If the credit line is not currently Defaulted.
+///
+/// # Events
+/// Emits a `("credit", "reinstate")` [`CreditLineEvent`].
 pub fn reinstate_credit_line(env: Env, borrower: Address) {
     require_admin_auth(&env);
 
@@ -201,6 +259,8 @@ pub fn reinstate_credit_line(env: Env, borrower: Address) {
 
     credit_line.status = CreditStatus::Active;
     env.storage().persistent().set(&borrower, &credit_line);
+    // Bump TTL: reinstated lines restart their active lifecycle.
+    bump_credit_line_ttl(&env, &borrower);
 
     publish_credit_line_event(
         &env,
