+++ rust-toolchain.toml
@@ -1,2 +1,2 @@
 [toolchain]
-channel = "stable-x86_64-pc-windows-msvc"
+channel = "stable"
